+++ example/CHANGELOG.md
@@ -1,5 +1,12 @@
 # example
 
+## 0.0.8
+
+### Patch Changes
+
+- Updated dependencies
+  - react-layered-dialog@1.0.0
+
 ## 0.0.7
 
 ### Patch Changes